python-cli/requirements.txt
PyYAML==6.0.2
jq==1.8.0
fastapi==0.115.5
jinja2 == 3.1.4